Sun, Aug 6, 2023, 1:16 AM +01:00

+++ readability/readability.c
@@ -114,8 +114,8 @@ int count_sentences(string text)
 
 int coleman_liau_formula(int letters,int words,int sentences)
 {
-    float L = (letters * 100) / words;
-    float S = (sentences * 100) / words;
+    float L = (letters * 100) / (float) words;
+    float S = (sentences * 100) / (float) words;
 
     float grade = lround(0.0588 * L - 0.296 * S - 15.8);
     return grade;
